requirements.txt
django==5.2.8
djangorestframework==3.15.2
python-dateutil==2.9.0.post0
